Refactor .gitignore and tsup.config.ts, and add package.json files
This commit refactors the .gitignore file in the Sample/TypeScript directory to include the \\node_modules\\ and \\dist\\ directories. It also updates the tsup.config.ts file to set the \\sourcemap\\ option to false. Additionally, it adds two new package.json files in the Sample/javascript and Sample/TypeScript directories.